requirements.txt
pandas>=0.24.1
python-decouple>=3.1
